requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# API clients
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
pytrends>=4.9.2

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# NLP & Sentiment Analysis
textblob>=0.17.1
nltk>=3.8.1
vaderSentiment>=3.3.2

# Content generation (optional but recommended)
google-generativeai>=0.3.0
openai>=1.3.0

# Utilities
pyyaml>=6.0.1
click>=8.1.7
colorama>=0.4.6
tqdm>=4.66.1

# Web Interface
flask>=2.3.0

# Date/time handling
python-dateutil>=2.8.2

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
responses>=0.23.0

# Development
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
